+++ R/sf.R
@@ -83,9 +83,9 @@ st_xy2sfc = function(x, as_points, ..., na.rm = TRUE) {
 	st_stars(x, dimensions = d)
 }
 
-st_as_sf.dimensions = function(x) {
+st_as_sf.dimensions = function(x, ...) {
 	ix = which_sfc(x)[1]
-	st_sf(setNames(list(x[[ ix ]]$values), names(x)[ix]), crs = st_crs(x))
+	st_sf(setNames(list(x[[ ix ]]$values), names(x)[ix]), crs = st_crs(x), ...)
 }
 
 
@@ -143,9 +143,9 @@ st_as_sf.stars = function(x, ..., as_points = FALSE, merge = FALSE, na.rm = TRUE
 			ex = attr(x[[1]], "exclude")
 			if(any(ex)){
 				id = (seq_along(ex) - 1)[!ex] # index-values matching to levels
-				ret[[1]] = structure(match(ret[[1]], id), class = "factor", levels = lev)
+				ret[[1]] = structure(match(ret[[1]], id), class = "factor", levels = lev) # ordered?
 			} else {
-				ret[[1]] = structure(ret[[1]], class = "factor", levels = lev)
+				ret[[1]] = structure(ret[[1]], class = "factor", levels = lev) # ordered?
 			}
 		}
 		st_set_crs(ret, crs)